requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
numpy==2.1.3
pillow==11.0.0
requests==2.32.3
